Fix timing problem, when read multiple servos
I added delayMicroseconds(800) in the GetCurrentDeg() one after digitalWrite(RS485_EN,LOW) for mimic the above serial print because when I print those out, the incoming data of servo 1 and 2 are not swapped, and get the correct value. And one more after OutputDeg = ReadReply() to not let the second servo read while the first one is reading. One more delayMicroseconds(2000) is added in the ReadReply(), to mimic the above serial print similar with the first delay I explained, so this delay and the first one make the reading command of multiple servo properly (without swapping value, data crash).